+++ src/main/java/net/bqc/aoc/year2023/Day03.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class Day03 extends Solution {
 
     @Override
-    public int solve(PART_NUMBER part, List<String> inputLines) {
+    public long solve(PART_NUMBER part, List<String> inputLines) {
         this.pup = part;
 
         int m = inputLines.size();
@@ -24,16 +24,16 @@ public class Day03 extends Solution {
             }
         }
 
-        int sum = 0;
+        long sum = 0;
         for (int row = 0; row < m; row++) {
             String line = inputLines.get(row);
             for (int col = 0; col < n; col++) {
                 char ch = line.charAt(col);
 
                 if (ch != '.' && !Character.isDigit(ch)) {
-                    List<Integer> partNumbers = findNearbyNumbers(matrix, m, n, row, col);
+                    List<Long> partNumbers = findNearbyNumbers(matrix, m, n, row, col);
                     if (this.pup == PART_NUMBER.ONE) {
-                        sum += partNumbers.stream().mapToInt(Integer::intValue).sum();
+                        sum += partNumbers.stream().mapToLong(Long::intValue).sum();
                     }
                     else if (this.pup == PART_NUMBER.TWO && ch == '*' && partNumbers.size() == 2) {
                         sum += partNumbers.get(0) * partNumbers.get(1);
@@ -46,37 +46,37 @@ public class Day03 extends Solution {
         return sum;
     }
 
-    private List<Integer> findNearbyNumbers(char[][] matrix, int m, int n, int row, int col) {
-        List<Integer> nearbyNumbers = new ArrayList<>();
+    private List<Long> findNearbyNumbers(char[][] matrix, int m, int n, int row, int col) {
+        List<Long> nearbyNumbers = new ArrayList<>();
 
         // Find nearby numbers in 8 directions
-        int up = findNumberAt(matrix, m, n, row - 1, col); // up
+        long up = findNumberAt(matrix, m, n, row - 1, col); // up
         nearbyNumbers.add(up);
         if (up == -1) {
-            int upLeft = findNumberAt(matrix, m, n, row - 1, col - 1); // up-left
-            int upRight = findNumberAt(matrix, m, n, row - 1, col + 1); // up-right
+            long upLeft = findNumberAt(matrix, m, n, row - 1, col - 1); // up-left
+            long upRight = findNumberAt(matrix, m, n, row - 1, col + 1); // up-right
             nearbyNumbers.add(upLeft);
             nearbyNumbers.add(upRight);
         }
 
-        int down = findNumberAt(matrix, m, n, row + 1, col); // down
+        long down = findNumberAt(matrix, m, n, row + 1, col); // down
         nearbyNumbers.add(down);
         if (down == -1) {
-            int downLeft = findNumberAt(matrix, m, n, row + 1, col - 1); // down-left
-            int downRight = findNumberAt(matrix, m, n, row + 1, col + 1); // down-right
+            long downLeft = findNumberAt(matrix, m, n, row + 1, col - 1); // down-left
+            long downRight = findNumberAt(matrix, m, n, row + 1, col + 1); // down-right
             nearbyNumbers.add(downLeft);
             nearbyNumbers.add(downRight);
         }
 
-        int left = findNumberAt(matrix, m, n, row, col - 1); // left
-        int right = findNumberAt(matrix, m, n, row, col + 1); // right
+        long left = findNumberAt(matrix, m, n, row, col - 1); // left
+        long right = findNumberAt(matrix, m, n, row, col + 1); // right
         nearbyNumbers.add(left);
         nearbyNumbers.add(right);
 
         return nearbyNumbers.stream().filter(x -> x > 0).toList();
     }
 
-    private int findNumberAt(char[][] matrix, int m, int n, int row, int col) {
+    private long findNumberAt(char[][] matrix, int m, int n, int row, int col) {
         if (row < 0 || row >= m || col < 0 || col >= n) {
             return -1;
         }
@@ -98,6 +98,6 @@ public class Day03 extends Solution {
             num.insert(0, matrix[row][p]);
         }
 
-        return Integer.parseInt(num.toString());
+        return Long.parseLong(num.toString());
     }
 }
